improve the stability of the viscous hydro simulation
In the extreme dilute region, the pi_size could be small negative (~-1e-32).
In these situations, the regulation of Wmunu tensor will not be
triggered because the if statement is not effective with a nan
rho_shear. Although the size of Wmunu is small at beginning, they can
potentially cumulate and eventually crashed the simulations.

Here we explicitly check whether rho_shear is nan or not before checking
its size. If rho_shear is negative, we will set Wmunu to 0. I tested the
improved version on a previously crashed event and the run was able to
cure the instability.