introduce a centrality dependent plateau width for PbPb collisions at LHC
With initial_profile == 112, the centrality dependence of particle production
is too flat compared with the experiments. Here we will vary the width
of the plateau of the longitudinal envelope profile according to the
total number of participants of the events. This modification right now
is only introduced at the LHC energies. We will only test PbPb
collisions for now as the total_number_of_nucleons is hard coded.